refactor : 폴더 위치 변경, 상수 데이터 분리
