+++ src/components/dialog/types.ts
@@ -1,0 +1,17 @@
+export interface Props {
+  searchInfo: {
+    tableId: string;
+    day?: string;
+    time?: number;
+  } | null;
+  onClose: () => void;
+}
+
+export interface SearchOption {
+  query?: string;
+  grades: number[];
+  days: string[];
+  times: number[];
+  majors: string[];
+  credits?: number;
+}
